+++ addons/imgui-godot/ImGuiGD.cs
@@ -2,75 +2,78 @@ using Godot;
 using ImGuiNET;
 using System;
 
-public static class ImGuiGD
+namespace ImGuiGodot
 {
-    public static float JoyAxisDeadZone { get; set; } = 0.15f;
-    public static bool JoyButtonSwapAB { get; set; } = false;
-
-    public static IntPtr BindTexture(Texture2D tex)
+    public static class ImGuiGD
     {
-        return ImGuiGDInternal.BindTexture(tex);
-    }
+        public static float JoyAxisDeadZone { get; set; } = 0.15f;
+        public static bool JoyButtonSwapAB { get; set; } = false;
 
-    public static void UnbindTexture(IntPtr texid)
-    {
-        ImGuiGDInternal.UnbindTexture(texid);
-    }
-
-    public static void Init()
-    {
-        if (IntPtr.Size != sizeof(ulong))
+        public static IntPtr BindTexture(Texture2D tex)
         {
-            GD.PrintErr("imgui-godot requires 64-bit pointers");
+            return ImGuiGDInternal.BindTexture(tex);
         }
-        ImGuiGDInternal.Init();
-    }
 
-    public static ImFontPtr AddFont(FontFile fontData, float fontSize, bool merge = false)
-    {
-        return ImGuiGDInternal.AddFont(fontData, fontSize, merge);
-    }
+        public static void UnbindTexture(IntPtr texid)
+        {
+            ImGuiGDInternal.UnbindTexture(texid);
+        }
 
-    public static ImFontPtr AddFont(FontFile fontData, float fontSize, IntPtr glyphRanges, bool merge = false)
-    {
-        return ImGuiGDInternal.AddFont(fontData, fontSize, glyphRanges, merge);
-    }
+        public static void Init()
+        {
+            if (IntPtr.Size != sizeof(ulong))
+            {
+                GD.PrintErr("imgui-godot requires 64-bit pointers");
+            }
+            ImGuiGDInternal.Init();
+        }
 
-    public static ImFontPtr AddFontDefault()
-    {
-        // right now this is a simple wrapper, but that will change
-        return ImGui.GetIO().Fonts.AddFontDefault();
-    }
+        public static ImFontPtr AddFont(FontFile fontData, float fontSize, bool merge = false)
+        {
+            return ImGuiGDInternal.AddFont(fontData, fontSize, merge);
+        }
 
-    // only call this once, shortly after Init
-    public static void RebuildFontAtlas()
-    {
-        ImGuiGDInternal.RebuildFontAtlas();
-    }
+        public static ImFontPtr AddFont(FontFile fontData, float fontSize, IntPtr glyphRanges, bool merge = false)
+        {
+            return ImGuiGDInternal.AddFont(fontData, fontSize, glyphRanges, merge);
+        }
 
-    public static void Update(double delta, Viewport vp)
-    {
-        ImGuiGDInternal.Update(delta, vp);
-    }
+        public static ImFontPtr AddFontDefault()
+        {
+            // right now this is a simple wrapper, but that will change
+            return ImGui.GetIO().Fonts.AddFontDefault();
+        }
 
-    public static void Render(RID parent)
-    {
-        ImGui.Render();
-        ImGuiGDInternal.RenderDrawData(ImGui.GetDrawData(), parent);
-    }
+        // only call this once, shortly after Init
+        public static void RebuildFontAtlas()
+        {
+            ImGuiGDInternal.RebuildFontAtlas();
+        }
 
-    public static void Shutdown()
-    {
-        ImGuiGDInternal.ClearCanvasItems();
-        if (ImGui.GetCurrentContext() != IntPtr.Zero)
-            ImGui.DestroyContext();
-    }
+        public static void Update(double delta, Viewport vp)
+        {
+            ImGuiGDInternal.Update(delta, vp);
+        }
 
-    /// <returns>
-    /// True if the InputEvent was consumed
-    /// </returns>
-    public static bool ProcessInput(InputEvent evt)
-    {
-        return ImGuiGDInternal.ProcessInput(evt);
+        public static void Render(RID parent)
+        {
+            ImGui.Render();
+            ImGuiGDInternal.RenderDrawData(ImGui.GetDrawData(), parent);
+        }
+
+        public static void Shutdown()
+        {
+            ImGuiGDInternal.ClearCanvasItems();
+            if (ImGui.GetCurrentContext() != IntPtr.Zero)
+                ImGui.DestroyContext();
+        }
+
+        /// <returns>
+        /// True if the InputEvent was consumed
+        /// </returns>
+        public static bool ProcessInput(InputEvent evt)
+        {
+            return ImGuiGDInternal.ProcessInput(evt);
+        }
     }
 }
